[REFACTOR] : 챌린지 관련 파일 관리
