+++ lib/layout.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:miyo/components/bottom_navbar.dart';
 import 'package:miyo/components/top_navbar.dart';
-import 'package:miyo/screens/challenge_screen.dart';
+import 'package:miyo/screens/challenges/challenge_screen.dart';
 import 'package:miyo/screens/imaginary_map_screen.dart';
 import 'package:miyo/screens/map_screen.dart';
 import 'package:miyo/screens/profile_screen.dart';
 
 class Layout extends StatefulWidget {
   final int initialIndex; // 초기 탭 인덱스를 받을 수 있도록 추가
-  
+
   const Layout({super.key, this.initialIndex = 0}); // 기본값은 0 (홈 탭)
 
   @override
@@ -46,10 +46,7 @@ class _LayoutState extends State<Layout> {
       appBar: TopNavbar(),
       // body 부분은 IndexedStack을 사용하여 선택된 페이지만 보여줍니다.
       // IndexedStack을 사용하면 각 페이지의 상태가 유지되므로 자연스러운 전환이 가능합니다.
-      body: IndexedStack(
-        index: _currentIndex,
-        children: _pages,
-      ),
+      body: IndexedStack(index: _currentIndex, children: _pages),
       bottomNavigationBar: BottomNavbar(
         currentIndex: _currentIndex,
         onTap: _onBottomNavTapped, // 이 콜백을 통해 선택된 인덱스 업데이트
