+++ lib/screens/challenges/challenge_detail_screen.dart
