+++ lib/screens/challenges/challenge_item.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:miyo/screens/challenge_detail_screen.dart';
+import 'package:miyo/screens/challenges/challenge_detail_screen.dart';
 
 class ChallengeItem extends StatelessWidget {
   // final IconData icon;
